CS410_TermProject_Group4-main/requirements.txt
bcrypt==4.1.3
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.8
iniconfig==2.0.0
inputimeout==1.0.4
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
keyring==25.2.1
more-itertools==10.3.0
MouseInfo==0.1.3
packaging==24.1
paramiko==3.4.0
pluggy==1.5.0
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyobjc-core==10.3.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.3.1; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.3.1; sys_platform == 'darwin'
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.2.2
pytest-mock==3.14.0
python-dotenv==1.0.1
keyring==25.2.1
pytweening==1.2.0
qasync==0.27.1
rubicon-objc==0.4.9
